Fix CSS not compiling on Platform.sh
Prior to this, the build steps on Platform.sh where failing,
leaving the site with no CSS. This was due to an error parsing
the regex pattern within the PurgeCSS extractor.

This has been fixed by changing the node version via nvm, and
including a `.environment` file as mentioned in https://docs.platform.sh/languages/nodejs/nvm.html.